requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-apikey==0.2.0
django-rest-auth==0.9.5
django-rest-authtoken==1.2.2
django-rest-knox==2.3.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
drf-jwt-knox==0.1.0
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==19.0.0
python3-openid==3.1.0
pytz==2018.7
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==3.3.1
